Move Maze::BugsnagConfig to Maze::Bugsnag::Config
